Platform Collision
I added platform types that can handle collisions and call a reset function that resets the triangle to position 0,0 where the level begins.

I also modified the gravity of the project, lowering it to make the speed increase slower.

Later I'd want to swap the rigid triangle for a kinematic body so I can handle the physics all on my own, but for now this is functional.